+++ src/gradient/linear.rs
@@ -4,15 +4,15 @@ use crate::{convert_colors, BlendMode, Color, Gradient, GradientBuilder, Gradien
 
 #[derive(Debug, Clone)]
 pub struct LinearGradient {
-    stops: Vec<(f64, [f64; 4])>,
-    domain: (f64, f64),
+    stops: Vec<(f32, [f32; 4])>,
+    domain: (f32, f32),
     mode: BlendMode,
     first_color: Color,
     last_color: Color,
 }
 
 impl LinearGradient {
-    pub(crate) fn new(colors: Vec<Color>, positions: Vec<f64>, mode: BlendMode) -> Self {
+    pub(crate) fn new(colors: Vec<Color>, positions: Vec<f32>, mode: BlendMode) -> Self {
         let dmin = positions[0];
         let dmax = positions[positions.len() - 1];
         let first_color = colors[0].clone();
@@ -33,7 +33,7 @@ impl LinearGradient {
 }
 
 impl Gradient for LinearGradient {
-    fn at(&self, t: f64) -> Color {
+    fn at(&self, t: f32) -> Color {
         if t <= self.domain.0 {
             return self.first_color.clone();
         }
@@ -81,7 +81,7 @@ impl Gradient for LinearGradient {
         }
     }
 
-    fn domain(&self) -> (f64, f64) {
+    fn domain(&self) -> (f32, f32) {
         self.domain
     }
 }
@@ -96,7 +96,7 @@ impl TryFrom<&GradientBuilder> for LinearGradient {
 }
 
 #[inline]
-fn linear_interpolation(a: &[f64; 4], b: &[f64; 4], t: f64) -> [f64; 4] {
+fn linear_interpolation(a: &[f32; 4], b: &[f32; 4], t: f32) -> [f32; 4] {
     [
         a[0] + t * (b[0] - a[0]),
         a[1] + t * (b[1] - a[1]),
@@ -106,7 +106,7 @@ fn linear_interpolation(a: &[f64; 4], b: &[f64; 4], t: f64) -> [f64; 4] {
 }
 
 #[inline]
-fn interp_angle(a0: f64, a1: f64, t: f64) -> f64 {
+fn interp_angle(a0: f32, a1: f32, t: f32) -> f32 {
     let delta = (((a1 - a0) % 360.0) + 540.0) % 360.0 - 180.0;
     (a0 + t * delta + 360.0) % 360.0
 }
